fix: cast loss to float32 before backward to prevent NaN gradients
When the model is loaded in FP16, outputs.loss is also float16. Calling
.backward() directly on a float16 scalar can produce NaN gradients due
to underflow in the gradient chain. Upcasting to float32 before backward
ensures stable gradient accumulation throughout Phase 1 calibration.

+++ src/salience/computer.py
@@ -139,7 +139,7 @@ class SalienceComputer:
                 loss = outputs.loss
 
                 if needs_grad and loss is not None:
-                    loss.backward()
+                    loss.float().backward()  # cast to float32 to prevent NaN gradients in FP16
 
                     if "hessian" in self.config.metrics:
                         self.hessian_metric.accumulate(self.model)
